Add feature3: gameboard with some logic
Added feature3 with a functioning gameboard that doesn't allow overriding of divs based on whether they don't already contain an X or O. Pair programming was a huge help. Rearranged the click event from just being in the app file to living in the events and refactored the events to be a simpler handler to be exported.